Suporte a download de videos mp4/mkv | audios mp3/m4a

+++ clearUploads.py
@@ -10,16 +10,19 @@ folders = [f for f in os.listdir(output_dir) if os.path.isdir(os.path.join(outpu
 
 # Loop pelas pastas
 for folder in folders:
+    # Se o nome da pasta for ".gitignore", pule para a próxima iteração
+    if folder == ".gitignore":
+        continue
 
-  # Obtem datetime de criação da pasta
-  folder_path = os.path.join(output_dir, folder)  
-  timestamp = os.path.getctime(folder_path)
-  creation_date = datetime.fromtimestamp(timestamp)
+    # Obtem datetime de criação da pasta
+    folder_path = os.path.join(output_dir, folder)  
+    timestamp = os.path.getctime(folder_path)
+    creation_date = datetime.fromtimestamp(timestamp)
 
-  # Limite de 30 minutos
-  limit = datetime.now() - timedelta(minutes=30) 
+    # Limite de 30 minutos
+    limit = datetime.now() - timedelta(minutes=30) 
 
-  # Se pasta foi criada a mais de 1 dia, deleta
-  if creation_date < limit:  
-    print(f'Deletando pasta antiga: {folder}')
-    shutil.rmtree(folder_path)
+    # Se pasta foi criada a mais de 30 minutos, deleta
+    if creation_date < limit:  
+        print(f'Deletando pasta antiga: {folder}')
+        shutil.rmtree(folder_path)
